add TProcessorFactory
